Move CacheConfig in config package
As another configuration class will be created, extract them in the same package.
