+++ src/main/java/com/orbitz/consul/cache/ServiceHealthCache.java
@@ -3,6 +3,7 @@ package com.orbitz.consul.cache;
 import com.google.common.net.HostAndPort;
 import com.google.common.primitives.Ints;
 import com.orbitz.consul.HealthClient;
+import com.orbitz.consul.config.CacheConfig;
 import com.orbitz.consul.model.health.ServiceHealth;
 import com.orbitz.consul.option.QueryOptions;
