+++ src/test/java/com/orbitz/consul/config/CacheConfigTest.java
@@ -1,8 +1,7 @@
-package com.orbitz.consul.cache;
+package com.orbitz.consul.config;
 
 import com.typesafe.config.Config;
 import com.typesafe.config.ConfigFactory;
-import org.junit.Assert;
 import org.junit.Test;
 
 import java.time.Duration;
